Adds admin override for public toggle
Allows administrators to toggle the public status of messages regardless of ownership.

This change introduces an `isAdmin` check to the `handleTogglePublic` function. If the user is an administrator, they can toggle the `isPublic` status even if they are not the owner of the message. This provides administrators with the ability to manage content visibility effectively.